refactor: Move `Capabilities` into own module.
