+++ openmls/src/treesync/node/leaf_node.rs
@@ -3,12 +3,11 @@ use openmls_traits::{
     key_store::OpenMlsKeyStore, signatures::Signer, types::Ciphersuite, OpenMlsCryptoProvider,
 };
 use serde::{Deserialize, Serialize};
+use thiserror::Error;
 use tls_codec::{
     Deserialize as TlsDeserializeTrait, Serialize as TlsSerializeTrait, TlsDeserialize,
     TlsSerialize, TlsSize, VLBytes,
 };
-
-use thiserror::Error;
 
 use crate::{
     binary_tree::array_representation::LeafNodeIndex,
@@ -18,8 +17,7 @@ use crate::{
     },
     credentials::{Credential, CredentialType, CredentialWithKey},
     error::LibraryError,
-    extensions::Extensions,
-    extensions::{Extension, ExtensionType, RequiredCapabilitiesExtension},
+    extensions::{Extension, ExtensionType, Extensions, RequiredCapabilitiesExtension},
     group::{config::CryptoConfig, GroupId},
     key_packages::KeyPackage,
     messages::proposals::ProposalType,
@@ -28,10 +26,13 @@ use crate::{
     versions::ProtocolVersion,
 };
 
-mod lifetime;
-pub use self::lifetime::Lifetime;
-
 use super::encryption_keys::{EncryptionKey, EncryptionKeyPair};
+
+mod capabilities;
+mod lifetime;
+
+pub use self::lifetime::Lifetime;
+pub use capabilities::*;
 
 #[derive(Error, Debug, PartialEq, Clone)]
 pub enum LeafNodeGenerationError<KeyStoreError> {
@@ -41,185 +42,6 @@ pub enum LeafNodeGenerationError<KeyStoreError> {
     /// Error storing leaf private key in key store.
     #[error("Error storing leaf private key in key store.")]
     KeyStoreError(KeyStoreError),
-}
-
-/// Capabilities of [`LeafNode`]s.
-///
-/// ```text
-/// struct {
-///     ProtocolVersion versions<V>;
-///     CipherSuite ciphersuites<V>;
-///     ExtensionType extensions<V>;
-///     ProposalType proposals<V>;
-///     CredentialType credentials<V>;
-/// } Capabilities;
-/// ```
-#[derive(
-    Debug, Clone, PartialEq, Eq, Serialize, Deserialize, TlsSerialize, TlsDeserialize, TlsSize,
-)]
-pub struct Capabilities {
-    versions: Vec<ProtocolVersion>,
-    ciphersuites: Vec<Ciphersuite>,
-    extensions: Vec<ExtensionType>,
-    proposals: Vec<ProposalType>,
-    credentials: Vec<CredentialType>,
-}
-
-// FIXME: deduplicate with CapabilitiesExtension.
-
-/// All extensions defined in the MLS spec are considered "default" by the spec.
-fn default_extensions() -> Vec<ExtensionType> {
-    vec![ExtensionType::ApplicationId]
-}
-
-/// All proposals defined in the MLS spec are considered "default" by the spec.
-fn default_proposals() -> Vec<ProposalType> {
-    vec![
-        ProposalType::Add,
-        ProposalType::Update,
-        ProposalType::Remove,
-        ProposalType::Presharedkey,
-        ProposalType::Reinit,
-        ProposalType::GroupContextExtensions,
-    ]
-}
-
-fn default_versions() -> Vec<ProtocolVersion> {
-    vec![ProtocolVersion::Mls10]
-}
-
-fn default_ciphersuites() -> Vec<Ciphersuite> {
-    vec![
-        Ciphersuite::MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519,
-        Ciphersuite::MLS_128_DHKEMP256_AES128GCM_SHA256_P256,
-        Ciphersuite::MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519,
-    ]
-}
-
-fn default_credentials() -> Vec<CredentialType> {
-    vec![CredentialType::Basic]
-}
-
-impl Capabilities {
-    /// Create new empty [`Capabilities`].
-    pub fn empty() -> Self {
-        Self {
-            versions: Vec::new(),
-            ciphersuites: Vec::new(),
-            extensions: Vec::new(),
-            proposals: Vec::new(),
-            credentials: Vec::new(),
-        }
-    }
-
-    /// Create a new [`Capabilities`] struct with the given configuration.
-    /// Any argument that is `None` is filled with the default values from the
-    /// global configuration.
-    pub fn new(
-        versions: Option<&[ProtocolVersion]>,
-        ciphersuites: Option<&[Ciphersuite]>,
-        extensions: Option<&[ExtensionType]>,
-        proposals: Option<&[ProposalType]>,
-        credentials: Option<&[CredentialType]>,
-    ) -> Self {
-        Self {
-            versions: match versions {
-                Some(v) => v.into(),
-                None => default_versions(),
-            },
-            ciphersuites: match ciphersuites {
-                Some(c) => c.into(),
-                None => default_ciphersuites(),
-            },
-            extensions: match extensions {
-                Some(e) => e.into(),
-                None => vec![],
-            },
-            proposals: match proposals {
-                Some(p) => p.into(),
-                None => vec![],
-            },
-            credentials: match credentials {
-                Some(c) => c.into(),
-                None => default_credentials(),
-            },
-        }
-    }
-
-    /// Get a reference to the list of versions in this extension.
-    pub fn versions(&self) -> &[ProtocolVersion] {
-        &self.versions
-    }
-
-    /// Set the versions list.
-    #[cfg(test)]
-    pub fn set_versions(&mut self, versions: Vec<ProtocolVersion>) {
-        self.versions = versions;
-    }
-
-    /// Get a reference to the list of ciphersuites in this extension.
-    pub fn ciphersuites(&self) -> &[Ciphersuite] {
-        &self.ciphersuites
-    }
-
-    /// Set the ciphersuites list.
-    #[cfg(test)]
-    pub fn set_ciphersuites(&mut self, ciphersuites: Vec<Ciphersuite>) {
-        self.ciphersuites = ciphersuites;
-    }
-
-    /// Get a reference to the list of supported extensions.
-    pub fn extensions(&self) -> &[ExtensionType] {
-        &self.extensions
-    }
-
-    /// Get a reference to the list of supported proposals.
-    pub fn proposals(&self) -> &[ProposalType] {
-        &self.proposals
-    }
-
-    /// Get a reference to the list of supported credential types.
-    pub fn credentials(&self) -> &[CredentialType] {
-        &self.credentials
-    }
-
-    /// Check if these [`Capabilities`] support all the capabilities
-    /// required by the given [`RequiredCapabilities`] extension. Returns
-    /// `true` if that is the case and `false` otherwise.
-    pub(crate) fn supports_required_capabilities(
-        &self,
-        required_capabilities: &RequiredCapabilitiesExtension,
-    ) -> bool {
-        // Check if all required extensions are supported.
-        if required_capabilities
-            .extension_types()
-            .iter()
-            .any(|e| !self.extensions().contains(e))
-        {
-            return false;
-        }
-        // Check if all required proposals are supported.
-        if required_capabilities
-            .proposal_types()
-            .iter()
-            .any(|p| !self.proposals().contains(p))
-        {
-            return false;
-        }
-        true
-    }
-}
-
-impl Default for Capabilities {
-    fn default() -> Self {
-        Capabilities {
-            versions: default_versions(),
-            ciphersuites: default_ciphersuites(),
-            extensions: default_extensions(),
-            proposals: default_proposals(),
-            credentials: default_credentials(),
-        }
-    }
 }
 
 pub type ParentHash = VLBytes;
@@ -797,13 +619,17 @@ impl LeafNode {
             .capabilities
             .extensions
             .contains(extension_type)
-            || default_extensions().iter().any(|et| et == extension_type)
+            || capabilities::default_extensions()
+                .iter()
+                .any(|et| et == extension_type)
     }
 
     /// Returns `true` if the [`ProposalType`] is supported by this leaf node.
     pub(crate) fn supports_proposal(&self, proposal_type: &ProposalType) -> bool {
         self.payload.capabilities.proposals.contains(proposal_type)
-            || default_proposals().iter().any(|pt| pt == proposal_type)
+            || capabilities::default_proposals()
+                .iter()
+                .any(|pt| pt == proposal_type)
     }
 
     /// Returns `true` if the [`CredentialType`] is supported by this leaf node.
